Separated MainPage jsx into 3 separate components
